Move JsLookupResult in a different file
